pyproject.toml
[project]
name = "smal-lang"
version = "0.2.3"
description = "State Machine Abstraction Language (SMAL) is a YAML-based description language for simple, robust and debuggable state machines."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz==0.21",
    "jinja2==3.1.6",
    "pydantic==2.12.5",
    "pyyaml==6.0.3",
    "semver==3.0.4",
    "typer==0.24.1",
]

[project.scripts]
smal = "smal.cli.app:app"

[dependency-groups]
build = [
    "uv==0.11.2",
]
dev = [
    "pytest==9.0.2",
    "pytest-cov==7.1.0",
    "python-semantic-release==9.21.1",
    "ruff==0.15.8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
smal = [
    "codegen/templates/*.j2",
    "codegen/templates/**/*.j2",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "conventional"
major_on_zero = false  # TODO: Change this when stable
allow_zero_version = true  # TODO: Remove this when major_on_zero == true
tag_format = "v{version}"
build_command = "uv build"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
exclude_commit_patterns = []
mode = "init"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix"]
other_allowed_tags = ["build", "chore", "ci", "docs", "style", "refactor", "test"]
allowed_tags = ["feat", "fix", "build", "chore", "ci", "docs", "style", "refactor", "test"]
default_bump_level = 0
parse_squash_commits = true
ignore_merge_commits = true

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

[tool.ruff]
target-version = "py310"
line-length = 160
select = ["ALL"]
ignore = [
    "ANN401", # Dynamically typed expressions (typing.Any) are disallowed in `**kwargs`

    "B011", # Do not `assert False` (`python -O` removes these calls), raise `AssertionError()`

    "FBT001", # Boolean-typed positional argument in function definition
    "FBT002", # Boolean default positional argument in function definition
    "FBT003", # Boolean positional value in function call

    "C901", # `foo` is too complex

    "FIX002", # Line contains TODO, consider resolving the issue

    "EM101", # Exception must not use a string literal, assign to variable first
    "EM102", # Exception must not use an f-string literal, assign to variable first

    "ERA001", # Found commented-out code [many false positives]

    "I001", # import block is un-sorted or un-formatted

    "INP001", # File is part of an implicit namespace package. Add an `__init__.py`.

    "LOG015", # `debug()` call on root logger

    "N802", # Function name should be lowercase
    "N803", # Argument name should be lowercase
    "N806", # Variable in function should be lowercase
    "N815", # Variable in class scope should not be mixedCase
    "N812", # Lowercase imported as non-lowercase

    "PLC0415", # `import` should be at the top-level of a file

    "PLR0911", # Too many return statements
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments in function definition
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used in comparison, consider replacing with a constant variable

    "PLW2901", # `for` loop variable overwritten by assignment target

    "PT011", # `pytest.raises(ValueError)` is too broad, set the `match` parameter or use a more specific exception

    "RET504", # Unnecessary assignment before `return` statement

    "S101", # Use of `assert` detected
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
    "S603", # `subprocess` call: check for execution of untrusted input
    "S607", # Starting a process with a partial executable path

    "SIM300", # Yoda condition detected (e.g., 3<x instead of x>3)
    "SIM108", # Use ternary operator `...` instead of `if`-`else`-block

    "T201", # `print` found
    "T203", # `pprint` found

    "TD002", # Missing author in TODO; try: `# TODO(<author_name>): ...` or `# TODO @<author_name>: ...`
    "TD003", # Missing issue link for this TODO

    "TRY003", # Avoid specifying long messages outside the exception class
    "TRY300", # Consider moving this statement to an `else` block
]

[tool.ruff.format]
quote-style = "double"
line-ending = "lf"